Separate module to define test classes
